[documenttranslation] generate container names
